third_party/blink/web_tests/virtual/fenced-frame-mparch/external/wpt/fenced-frame/disallowed-navigations-dangling-markup-urn.https-expected.txt

This is a testharness.js-based test.
[FAIL] fenced frame opaque URN => https: URL with dangling markup 'blo\nck<ed'
  assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/fenced-frame/resources/report-url.html?block<ed="
[FAIL] fenced frame opaque URN => https: URL with dangling markup 'blo\rck<ed'
  assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/fenced-frame/resources/report-url.html?block<ed="
[FAIL] fenced frame opaque URN => https: URL with dangling markup 'blo	ck<ed'
  assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/fenced-frame/resources/report-url.html?block<ed="
[FAIL] fenced frame opaque URN => https: URL with dangling markup 'blo<ck\ned'
  assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/fenced-frame/resources/report-url.html?blo<cked="
[FAIL] fenced frame opaque URN => https: URL with dangling markup 'blo<ck\red'
  assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/fenced-frame/resources/report-url.html?blo<cked="
[FAIL] fenced frame opaque URN => https: URL with dangling markup 'blo<ck	ed'
  assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/fenced-frame/resources/report-url.html?blo<cked="
Harness: the test ran to completion.
